Added StringSubstringIndex, StringSubstringLength commands (#335)
* Started on basic string substring properties

* Added StringSubstringIndex, StringSubstringLength commands

Fixes #330.

As per the issue: some languages don't support one of the two string substring methods. If a primitive is given as a second parameter for a variant not natively supported by the language, it'll become something like .substring(3, 7 - 3). GLS should natively parse that into substring(3, 4).
This is done with a new `MathResolver`, which can resolve addition and subtraction types for now.

Introduces #334 as a new roundtripping requirement.